+++ direwolf.c
@@ -231,13 +231,16 @@ int main (int argc, char *argv[])
 	text_color_init(t_opt);
 	text_color_set(DW_COLOR_INFO);
 	//dw_printf ("Dire Wolf version %d.%d (%s) Beta Test\n", MAJOR_VERSION, MINOR_VERSION, __DATE__);
-	dw_printf ("Dire Wolf DEVELOPMENT version %d.%d %s (%s)\n", MAJOR_VERSION, MINOR_VERSION, "H", __DATE__);
+	dw_printf ("Dire Wolf DEVELOPMENT version %d.%d %s (%s)\n", MAJOR_VERSION, MINOR_VERSION, "I", __DATE__);
 	//dw_printf ("Dire Wolf version %d.%d\n", MAJOR_VERSION, MINOR_VERSION);
 
-#if defined(ENABLE_GPSD) 	// later or hamlib ...
+#if defined(ENABLE_GPSD) || defined(USE_HAMLIB)
 	dw_printf ("Includes optional support for: ");
 #if defined(ENABLE_GPSD)
 	dw_printf (" gpsd");
+#endif
+#if defined(USE_HAMLIB)
+	dw_printf (" hamlib");
 #endif
 	dw_printf ("\n");
 #endif